% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataVisApp.R
\name{dataVis}
\alias{dataVis}
\title{dataVisR}
\usage{
dataVis(df)
}
\arguments{
\item{df}{A data frame to visualize.}
}
\value{
starts a shiny-app (no return value)
}
\description{
Creates a Shiny app that displays for each variable in the given data frame:
\itemize{
\item Summary statistics
\item Min/Max (for numeric variables)
\item Median or "No median" (if no repeating values)
\item Histogram (small variance of Integers) or Density plot (numeric) or barplot (categorical)
}
}
\examples{
if (interactive()) {
  dataVis(mtcars)
}
}
