% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signify.R
\name{signify}
\alias{signify}
\title{Makes your data significant}
\usage{
signify(df, cat_var, num_var, alpha = 0.05)
}
\arguments{
\item{df}{data.frame containig your sad, unsignificant data}

\item{cat_var}{name of the categorial variable (string)}

\item{num_var}{name of the numeric variable (string)}

\item{alpha}{signifiance level (by default: 0.05)}
}
\value{
data.frame with modified numeric variable
}
\description{
transforms a numeric variable so that t-tests on it are as significant as you desire
}
\examples{
iris_sign <- signify(iris, "Species", "Sepal.Width", 0.001)
}
